Removed angular result functionality and moved server functionality to the cloud folder
