--- lib/angular/result-ng.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = 'wfm.result';
-
-angular.module('wfm.result', [
-  require('./service')
-]);
