+++ lib/cloud/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var config = require('./config');
+var config = require('../config');
 var shortid = require('shortid');
 var ResultTopics = require('fh-wfm-mediator/lib/topics');
