+++ package.json
@@ -2,7 +2,6 @@
   "name": "fh-wfm-result",
   "version": "0.1.2",
   "description": "A result module for WFM, for working with the results of pushing a workorder through a workflow",
-  "main": "lib/angular/result-ng.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/feedhenry-raincatcher/raincatcher-result.git"
